requirement.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.10.4
async-timeout==5.0.1
asyncio-throttle==1.0.2
attrs==25.3.0
backports.zoneinfo==0.2.1
black==24.8.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
cryptography==45.0.7
dashscope==1.24.2
distro==1.9.0
exceptiongroup==1.3.0
fake-useragent==1.4.0
flake8==7.1.2
frozenlist==1.5.0
greenlet==3.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.1.0
isort==5.13.2
jiter==0.9.1
loguru==0.7.2
mccabe==0.7.0
multidict==6.1.0
mypy==1.14.1
mypy_extensions==1.1.0
openai==1.105.0
packaging==25.0
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.3.6
playwright==1.40.0
pluggy==1.5.0
propcache==0.2.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
pyflakes==3.2.0
pytest==8.3.5
pytest-asyncio==0.24.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
ruff==0.12.11
six==1.17.0
sniffio==1.3.1
tenacity==8.2.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.3
uv==0.8.15
websocket-client==1.8.0
win32_setctime==1.2.0
-e git+https://github.com/heheshang/auto-publish.git@1f2eb52724616495b3d3e21ae3fafca2907876a8#egg=xiaohongshu_auto_publisher
yarl==1.15.2
zipp==3.20.2
